refactor: use int for entries' ids

+++ Assets/Source/Script/Core/CharacterEntry.cs
@@ -5,13 +5,13 @@
         public string Text { get; }
         public Character CharacterData { get; }
         
-        public CharacterEntry(ulong id, float duration, string text, Character character) : base(id, duration)
+        public CharacterEntry(int id, float duration, string text, Character character) : base(id, duration)
         {
             Text = text;
             CharacterData = character;
         }
 
-        public CharacterEntry(ulong id, float duration, ExpireEvent expireEvent, string text, Character character) : 
+        public CharacterEntry(int id, float duration, ExpireEvent expireEvent, string text, Character character) : 
             base(id, duration, expireEvent)
         {
             Text = text;
